requirements_db.txt
# PostgreSQL数据库依赖
SQLModel==0.0.14
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

# 其他必要依赖
pydantic==2.5.2
python-dotenv==1.0.0  # 用于从.env文件加载环境变量

# 开发调试依赖
uvicorn==0.24.0.post1
fastapi==0.104.1